demo/requirements.txt
certifi==2020.4.5.1
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
